refactor: make pluggable online executor for simple cycle
