+++ autora/cycle/simple.py
@@ -1,67 +1,10 @@
-import copy
-from collections.abc import Mapping
-from dataclasses import dataclass, replace
-from typing import Callable, Dict, Iterable, List, Optional
+from __future__ import annotations
 
-import numpy as np
-from sklearn.base import BaseEstimator
+from typing import Callable, Dict, Optional
 
-from autora.experimentalist.pipeline import Pipeline
-from autora.utils.dictionary import LazyDict
+from autora.cycle._executor import FullCycleExecutor
+from autora.cycle.state import SimpleCycleData
 from autora.variable import VariableCollection
-
-
-@dataclass(frozen=True)
-class SimpleCycleData:
-    """An object passed between and updated by processing steps in the SimpleCycle."""
-
-    # Static
-    metadata: VariableCollection
-
-    # Aggregates each cycle from the:
-    # ... Experimentalist
-    conditions: List[np.ndarray]
-    # ... Experiment Runner
-    observations: List[np.ndarray]
-    # ... Theorist
-    theories: List[BaseEstimator]
-
-
-def _get_cycle_properties(data: SimpleCycleData):
-    """
-    Examples:
-        Even with an empty data object, we can initialize the dictionary,
-        >>> cycle_properties = _get_cycle_properties(SimpleCycleData(metadata=VariableCollection(),
-        ...     conditions=[], observations=[], theories=[]))
-
-        ... but it will raise an exception if a value isn't yet available when we try to use it
-        >>> cycle_properties["%theories[-1]%"] # doctest: +ELLIPSIS
-        Traceback (most recent call last):
-        ...
-        IndexError: list index out of range
-
-        Nevertheless, we can iterate through its keys no problem:
-        >>> [key for key in cycle_properties.keys()] # doctest: +NORMALIZE_WHITESPACE
-        ['%observations.ivs[-1]%', '%observations.dvs[-1]%', '%observations.ivs%',
-        '%observations.dvs%', '%theories[-1]%', '%theories%']
-
-    """
-
-    n_ivs = len(data.metadata.independent_variables)
-    n_dvs = len(data.metadata.dependent_variables)
-    cycle_property_dict = LazyDict(
-        {
-            "%observations.ivs[-1]%": lambda: data.observations[-1][:, 0:n_ivs],
-            "%observations.dvs[-1]%": lambda: data.observations[-1][:, n_ivs:],
-            "%observations.ivs%": lambda: np.row_stack(
-                [np.empty([0, n_ivs + n_dvs])] + data.observations
-            )[:, 0:n_ivs],
-            "%observations.dvs%": lambda: np.row_stack(data.observations)[:, n_ivs:],
-            "%theories[-1]%": lambda: data.theories[-1],
-            "%theories%": lambda: data.theories,
-        }
-    )
-    return cycle_property_dict
 
 
 class SimpleCycle:
@@ -327,6 +270,7 @@ class SimpleCycle:
         [array([2]), array([6]), array([5]), array([7]), array([3]), array([4])]
 
         By using the monitor callback, we can investigate what's going on with the cycle properties:
+        >>> from autora.cycle.params import _get_cycle_properties
         >>> cycle_with_cycle_properties.monitor = lambda data: print(
         ...     _get_cycle_properties(data)["%observations.ivs%"].flatten()
         ... )
@@ -364,6 +308,7 @@ class SimpleCycle:
         experiment_runner,
         monitor: Optional[Callable[[SimpleCycleData], None]] = None,
         params: Optional[Dict] = None,
+        executor_collection=FullCycleExecutor,
     ):
         """
         Args:
@@ -380,9 +325,12 @@ class SimpleCycle:
                 `{"experimentalist": {"pool": {"n": 30}}}`
         """
 
-        self.theorist = theorist
-        self.experimentalist = experimentalist
-        self.experiment_runner = experiment_runner
+        self.executor_collection = executor_collection(
+            theorist_estimator=theorist,
+            experiment_runner_callable=experiment_runner,
+            experimentalist_pipeline=experimentalist,
+        )
+
         self.monitor = monitor
         if params is None:
             params = dict()
@@ -401,127 +349,14 @@ class SimpleCycle:
         return self
 
     def __next__(self):
-        assert (
-            "experiment_runner" not in self.params
-        ), "experiment_runner cannot yet accept cycle properties"
-        assert (
-            "theorist" not in self.params
-        ), "theorist cannot yet accept cycle properties"
-
-        data = self.data
-        params_with_cycle_properties = _resolve_cycle_properties(
-            self.params, _get_cycle_properties(self.data)
-        )
-
-        data = self._experimentalist_callback(
-            self.experimentalist,
-            data,
-            params_with_cycle_properties.get("experimentalist", dict()),
-        )
-        data = self._experiment_runner_callback(self.experiment_runner, data)
-        data = self._theorist_callback(self.theorist, data)
+        data = self.executor_collection.full_cycle(self.data, self.params)
         self._monitor_callback(data)
         self.data = data
-
         return self
 
     def __iter__(self):
         return self
 
-    @staticmethod
-    def _experimentalist_callback(
-        experimentalist: Pipeline, data_in: SimpleCycleData, params: dict
-    ):
-        new_conditions = experimentalist(**params)
-        if isinstance(new_conditions, Iterable):
-            # If the pipeline gives us an iterable, we need to make it into a concrete array.
-            # We can't move this logic to the Pipeline, because the pipeline doesn't know whether
-            # it's within another pipeline and whether it should convert the iterable to a
-            # concrete array.
-            new_conditions_values = list(new_conditions)
-            new_conditions_array = np.array(new_conditions_values)
-        else:
-            raise NotImplementedError(f"Object {new_conditions} can't be handled yet.")
-
-        assert isinstance(
-            new_conditions_array, np.ndarray
-        )  # Check the object is bounded
-        data_out = replace(
-            data_in,
-            conditions=data_in.conditions + [new_conditions_array],
-        )
-        return data_out
-
-    @staticmethod
-    def _experiment_runner_callback(
-        experiment_runner: Callable, data_in: SimpleCycleData
-    ):
-        x = data_in.conditions[-1]
-        y = experiment_runner(x)
-        new_observations = np.column_stack([x, y])
-        data_out = replace(
-            data_in, observations=data_in.observations + [new_observations]
-        )
-        return data_out
-
-    @staticmethod
-    def _theorist_callback(theorist, data_in: SimpleCycleData):
-        all_observations = np.row_stack(data_in.observations)
-        n_xs = len(
-            data_in.metadata.independent_variables
-        )  # The number of independent variables
-        x, y = all_observations[:, :n_xs], all_observations[:, n_xs:]
-        if y.shape[1] == 1:
-            y = y.ravel()
-        new_theorist = copy.deepcopy(theorist)
-        new_theorist.fit(x, y)
-        data_out = replace(
-            data_in,
-            theories=data_in.theories + [new_theorist],
-        )
-        return data_out
-
     def _monitor_callback(self, data: SimpleCycleData):
         if self.monitor is not None:
             self.monitor(data)
-
-
-def _resolve_cycle_properties(params: Dict, cycle_properties: Mapping):
-    """
-    Resolve "cycle properties" inside a nested dictionary.
-
-    In this context, a "cycle property" is a string which is meant to be replaced by a
-    different value before the dictionary is used.
-
-    Args:
-        params: a (nested) dictionary of keys and values, where some values might be
-            "cycle property names"
-        cycle_properties: a dictionary of "cycle property names" and their "real values"
-
-    Returns: a (nested) dictionary where "cycle property names" are replaced by the "real values"
-
-    Examples:
-
-        >>> params_0 = {"key": "%foo%"}
-        >>> cycle_properties_0 = {"%foo%": 180}
-        >>> _resolve_cycle_properties(params_0, cycle_properties_0)
-        {'key': 180}
-
-        >>> params_1 = {"key": "%bar%", "nested_dict": {"inner_key": "%foobar%"}}
-        >>> cycle_properties_1 = {"%bar%": 1, "%foobar%": 2}
-        >>> _resolve_cycle_properties(params_1, cycle_properties_1)
-        {'key': 1, 'nested_dict': {'inner_key': 2}}
-
-    """
-    params_ = copy.copy(params)
-    for key, value in params_.items():
-        if isinstance(value, dict):
-            params_[key] = _resolve_cycle_properties(value, cycle_properties)
-        elif (
-            isinstance(value, str) and value in cycle_properties
-        ):  # value is a key in the cycle_properties dictionary
-            params_[key] = cycle_properties[value]
-        else:
-            pass  # no change needed
-
-    return params_
